+++ pkg/views/provider/install/select.go
@@ -1,7 +1,7 @@
 // Copyright 2024 Daytona Platforms Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package provider_install
+package install
 
 import (
 	"sort"
